requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
Django==4.2
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
python-telegram-bot==21.9
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
yarl==1.18.3
